app/requirements.txt
streamlit
sentence-transformers
faiss-cpu
openai
langchain-text-splitters
pypdf
tiktoken
